Migrate from SQL to Mongo (#10)
I had an opportunity to use this data set for a class, but one of the requirements for the project was MongoDB. I decided to just replace the existing SQL Pipeline with a MongoDB Pipeline. This transition required the rework/simplification of the Items in the project, as well as extensive updates to testing data.

The notable changes to Items are that there are now only 2 of them: Area and Route. This was done because I felt that an embedded data model would suit MongoDB better. On top of that, because the size of the conditions and grade data is finite (max 12 and 7 respectively), there is no concern about documents getting bloated.