Change parent_id to ancestor_ids (#17)
While working on building aggregations I realized that the parent linked tree structure I had set up was inadequate for some of the things I wanted to do. Specifically per-area route metadata statistics (eg. how many of a specific type of route are in an area) would be prohibitively difficult (at least with my current knowledge of MongoDB).

To address this, I replaced the parent_id field with an ancestors field which contains an ordered list of a given resource's ancestors where the 0th index is the topmost parent. I chose to exclude the default topmost resource (the route guide) because at that point I'm just searching the entire dataset so I could perform a query on the route or area collection as a whole.